+++ src/Lwt/Models/TermReadModel.cs
@@ -20,7 +20,7 @@ namespace Lwt.Models
         /// </summary>
         public LearningLevel LearningLevel { get; set; } = null!;
 
-        public int Index { get; set; }
+        public long Index { get; set; }
 
         public string? Meaning { get; set; }
