+++ src/Lwt/Models/TextTerm.cs
@@ -32,6 +32,6 @@ namespace Lwt.Models
         /// <summary>
         /// Gets or sets index.
         /// </summary>
-        public int Index { get; set; }
+        public long Index { get; set; }
     }
 }
