+++ src/Lwt/Services/TextService.cs
@@ -270,9 +270,9 @@ namespace Lwt.Services
                 .CountAsync();
         }
 
-        public async Task<int> GetProcessedTermCountAsync(int id, int userId)
+        public async Task<long> GetProcessedTermCountAsync(int id, int userId)
         {
-            int? count = await this.textRepository.Queryable()
+            long? count = await this.textRepository.Queryable()
                 .Where(t => t.Id == id && t.UserId == userId)
                 .Select(t => t.ProcessedTermCount)
                 .FirstOrDefaultAsync();
